Factor out the blitter sequencer.
